pyproject.toml
[project]
name = "fastapi-infra-kit"
version = "1.0.0"
description = "A FastAPI infrastructure kit for building robust applications"
readme = "README.md"
requires-python = "==3.12.1"
dependencies = [
    "alembic==1.16.4",
    "asyncpg>=0.30.0",
    "fastapi-pagination>=0.13.3",
    "fastapi[standard-no-fastapi-cloud-cli]==0.116.1",
    "isort>=6.0.1",
    "loguru>=0.7.3",
    "psycopg2-binary==2.9.10",
    "pydantic-settings==2.10.1",
    "python-json-logger==3.3.0",
    "python-multipart==0.0.20",
    "ruff==0.12.7",
    "sqlakeyset>=2.0.1746777265",
    "sqlalchemy==2.0.42",
]
